requirements.txt
Django
djangorestframework
django-cors-headers
mysqlclient
python-dotenv
pymilvus
requests
celery
redis
django-celery-results
sentence-transformers
djangorestframework-simplejwt
openai
langchain
langchain_huggingface
langchain_openai
pymupdf 
markdownify
langchain_text_splitters
langchain_community
pymysql
channels-redis
channels
py7zr
beautifulsoup4
lxml